feat; clean up types
