+++ src/createBrowserRouter.tsx
@@ -1,13 +1,24 @@
 import BrowserProtocol from 'farce/BrowserProtocol';
 import React from 'react';
 
-import createFarceRouter from './createFarceRouter';
-import resolver from './resolver';
-import {
-  BrowserRouter,
-  BrowserRouterOptions,
+import { RenderArgs } from './ElementsRenderer';
+import createFarceRouter, {
+  FarceRouterOptions,
   FarceRouterProps,
-} from './typeUtils';
+} from './createFarceRouter';
+import resolver from './resolver';
+import { Resolver } from './typeUtils';
+
+export interface BrowserRouterProps
+  extends Omit<FarceRouterProps, 'resolver'> {
+  resolver?: Resolver;
+}
+
+export type BrowserRouter = React.ComponentType<BrowserRouterProps>;
+export interface BrowserRouterOptions
+  extends Omit<FarceRouterOptions, 'historyProtocol'> {
+  render?: (args: RenderArgs) => React.ReactElement;
+}
 
 export default function createBrowserRouter(
   options: BrowserRouterOptions,
